[REFACTOR] 테스트 환경 리팩토링 - 더미데이터 파일명을 get-label-list에서 getLabelList로 변경했습니다. - setUpTest파일에 몰려있던 rest handler을 분리하였습니다.
